SBS/requirements.txt
confusable-homoglyphs==3.2.0
Django==2.1.1
django-appconf==1.0.2
django-axes==4.4.0
django-filter==2.0.0
django-guardian==1.4.9
django-ipware==2.1.0
django-otp==0.5.0
django-session-security==2.6.1
django-su==0.8.0
mysqlclient==1.3.13
pycryptodome==3.6.6
python-dotenv==0.9.1
pytz==2018.5
qrcode==6.0
six==1.11.0
